requirements.txt
fastapi == 0.68.0
uvicorn == 0.15.0
PyYAML == 5.4.1
APScheduler == 3.7.0
loguru == 0.5.3
aiohttp ==  3.7.4
pylint 